api/requirements.txt
pydantic~=2.10.6
python-dotenv~=1.0.1
httpx~=0.28.1
fastapi~=0.115.7
loguru~=0.7.3
motor~=3.7.0
pandas~=2.2.3
APScheduler~=3.11.0
uvicorn~=0.34.0
pytz==2025.1
passlib[bcrypt]
python-jose[cryptography]